build: Use setup-go action cache.
The latest version of the setup-go action now caches the go module and
build cache by default, so it should no longer be done separately to
avoid saving the same data under two different cache keys.

This updates the caching logic accordingly to remove those directories
from the separate cache step and renames the step to reflect that it now
only involves the lint cache.

Finally, while here, it also switches the key over to use the hash of
the 'go.sum' file in the root of the repository as opposed to using the
commit hash as it produces significantly less cache objects.