[release-v1.8] rpcserver: Modify getnetworkhashps -1 blocks logic.
Now that the target difficulty is calculated every block as opposed to
on an interval, the logic in the getnetworkhashps RPC that treated a
negative number of blocks as a signal to calculate back to the most
recent difficulty change no longer makes sense.

Thus, this updates the getnetworkhashps RPC to ignore negative values
so that the default number of blocks is used instead.

It also updates the RPC server help description accordingly to remove
mention of the previous behavior.