rpcserver: Improve internal error handling.
This cleans up the handling of internal errors in the RPC server to be
more consistent and use the errors themselves as opposed to their
stringized form.

There are far more instances of already having an error to pass along
versus the ones that generate new errors at the call site.  Accepting
the errors directly results in less noise and also makes it harder to
misuse.