server: Use server peer in log statements.
This change a few cases that log the peer to use the server peer instead
of the underlying peer from the peer package.  This doesn't change the
output in any way, but it is more consistent with the vast majority of
other instances.  It also preferable since it ensures consistent logging
output if a custom stringer is ever implemented on server peers.