gcs: Add func to determine max cfilter size.
This will be helpful in determining the max number of cfilters to return
in a future batched getcfilter message.

This will also nail down the maximum size of the filter based on the
constants used throughout the project, such that if any of them changes,
causing the max cfilter size to change, a test will break indicating the
need to review any code paths that assume a max cfilter size.