wire: Add msgs to get batch of cfilters.
This adds the getcfsv2 and cfiltersv2 messages. These messages are
intended to allow nodes in the P2P network to request and receive a
batch of version 2 committed filters that span multiple sequential
blocks in a chain.

These messages are intended to be used when syncing SPV clients, which
require requesting all committed filters on the current main chain and
currently use a large number of getcfilter/cfilter messages.

One of the critical issues in the design for these messages is the
MaxCFiltersV2PerBatch constant, which establishes an upper bound on the
max number of cfilters requested and replied.  The current value of this
constant was decided based on the max block size for all of the
currently deployed networks, the max possible filter size for
transactions in such a block and the max P2P message size.

A runtime check is added to ensure any changes to the constants that
were involved in deciding this number trigger a panic so that this bound
is verified and, if needed, a new protocol version is introduced to
update it.  This check is meant to guard against inadvertedly changing
the assumptions that went into establishing this bound without reviewing
it.