dcrutil: Use os.UserHomedir for base home directory in tests
The home directory associated with a user (e.g. their passwd entry on Unix)
may not match the HOME environment variable that the non-test code will use
for the home directory.  There's no reason to fail the test if these differ,
so use os.UserHomeDir across the board.

Observed in a sandboxed nix build environment where the user's passwd entry
used /build for the home directory, but $HOME was set to /homeless-shelter.