wire: Add epoch field to mix key exchange message
The epoch will be included in the derivation of the session hash, instead of
only deriving a session from previous pair requests.  This allows new messages
in following epochs to not conflict with previous session attempts of the same
pair request set.

This also provides a decent heuristic to discover and exclude any peers that
were unresponsive for a full epoch or more.

This is a breaking API change but the wire module has not seen a release since
mixing messages were added.