rpcserver: Remove unneeded AddedNodeInfo method.
The RPC server was refactored some time ago to decouple it from the
internals of the main server and part of that refactoring modified the
RPC server to make use of a local peer interface to obtain all necessary
information about peers which is then used to create RPC server results.
Due to that, the AddedNodeInfo and PersistentPeers methods now do the
exact same thing.

This modifies the rpcserver.ConnManager interface to remove the now
unnecessary AddedNodeInfo method in favor of the PersistentPeers method
and updates all relevant code accordingly.