server: Use atomic types for some svr peer fields.
This updates a couple of cases of server peer data fields protected by
an individual mutex to make use of the newer atomic types introduced in
Go 1.19 instead.

This is desirable since atomic types are both simpler to use and more
efficient when the data being protected does not require additional
logic in the critical section as is the case for the updated fields.