connmgr: Only mark persistent peer reconn pending.
The current code is incorrectly treating non-persistent peer requests as
though they will be reconnected to whenever the target number of
outbound peers has not been reached by marking them as pending and
logging the reconnection attempt that will never come.  In addition to
the spurious logging, it also means that manually attempting to
reconnect to such a peer will incorrectly believe a connection is
already pending.

This resolves that issue by only logging the attempt and adding the
connection request back to the pending map for persistent peers.