mining: Update blk templ diff for too few voters.
This updates the block template generation path for handling too few
voters to also update the required difficulty bits in the created block
template since the new ASERT difficulty algorithm changes every block
and is influenced by the timestamp.

This is necessary to prevent the possibility of generating block
templates that fail to validate and have to be discarded resulting in
unnecessary additional work while aggressive mining in the
aforementioned case of not having enough voters.