cpuminer: Rework discrete mining vote wait logic.
In an effort to provide a better user experience for discrete mining
which is only used in testing scenarios, it contains additional logic
that ultimately aims to wait for votes to show up prior to mining a new
block when it is invoked in rapid succession as opposed to just
generating a bunch of side chain block which would otherwise happen.

The current logic attempts to accomplish that goal by storing the parent
hash of the block the discrete mining process most recently mined and
waiting for a new template if invoked again so long as the current block
template continues to build on that stored parent hash.

Unfortunately, there are various scenarios where you still would want to
generate a new block that builds on the same parent hash such as the
case when the current tip is manually invalidated or when the current
tip is unable to get enough votes to be built on.

Rather than trying to identify and special case those scenarios, this
reworks the logic to instead just store a pointer to the template that
the discrete mining process most recently mined instead and then wait
for a new template if invoked again while the same template is still
active.

This approach is much more robust and also provides additional
flexibility such as allowing the caller to override the wait by manually
regenerating a new template.