wire: Add Opcode field to MixPairReqUTXO
Validating contributed signatures of vote, revocation, and treasury generation
outputs requires knowing the exact opcode tag used by the previous output
script, as it is committed to by the sighash.  Include it in the
MixPairReqUTXO so that all clients are able to validate the merged signatures
in the coinjoin.  The special value 0 (OP_FALSE) will be used for regular tree
outputs that do not use a no-op tag.

This is a breaking API change but the wire module has not seen a release since
mixing messages were added.