wire: Remove PR hashes from (get)initstate
This reverts all of the initstate changes made in
b9d8d49c901bb7cbb19ed36d636c3e3d86a1fe43.

The limit of 100 was too too low to be useful and cranking this value up was
not a proper solution either.  It also only synced pair request messages,
which would cause mid-run mixes to leave stale PRs around in the mixpool until
they expired.

This is handled better in newer mixpool code by requesting unknown pair
request messages when they are seen in a key exchange.  The mixpool records
the KE as an orphan and will process it after all missing PRs have been
requested and processed.