wire: add flag bits to PR message
Two new flag fields are introduced to the pair request message, both of which
are application specific and have no meaning to the wire package.  The first
is a Flags field which will be used to describe capabilities that do not
affect the pairing type.  This will be used in the initial release to allow
wallets to advertise support for solving and publishing the factored
polynomial.  While here, a second flags field which affects the pairing
compatibility is also added, as this will aid in any future incompatibility in
the client behavior without disturbing clients who have not upgraded.