server: Implement separate mutex for peer state.
Currently, all operations related to adding, removing, and banning peers
are implemented via a single goroutine and channels to protect
concurrent access.

The existing implementation has worked well for over a decade, however,
it is really not ideal in a few ways:

- It is difficult to improve parallel processing since everything is
  forced into a single goroutine
- It requires a lot of plumbing to provide access to any related
  information
- The use of multiple channels means the ordering of events is not
  entirely well defined.

In regards to the final point about event ordering, one example is that
it's possible for a peer to be removed before it was ever added.  The
surrounding code is aware of these possibilities and handles things
gracefully, but it really is not ideal.

In practice, channels are best suited for passing ownership of data,
distributing units of work, and communicating async results while
mutexes are better suited for caches and state.

Converting all of the code the code related to updating and querying the
server's peer state to synchronous code that makes use of a separate
mutex to protect it will address the aforementioned concerns.  Namely,
it:

- Improves the semantics in regards to the aforementioned ordering
- Ultimately allows more code to run in parallel in the individual peer
  goroutines
- Requires less plumbing for updating and querying the state
- Makes the state available to calling code so it can make better
  decisions

Thus, this is part of a series of commits that aims to make that
conversion.

This first commit introduces a separate mutex on the peer state and
updates the code to protect all references to the relevant fields with
that mutex.  This will allow future commits to methodically refactor the
various operations without introducing races.