server: Make server peer conn req concurrent safe.
This updates the server peer connection request field to be safe for
concurrent access by making it an atomic pointer.

It also reorders some of the server peer fields and adds some comments
that call out the concurrency semantics while here.

This is a part of the overall effort to convert all of the code related
to updating and querying the server's peer state to synchronous code
that makes use of a separate mutex to protect it.