server: Make done peer logic synchronous.
This refactors the logic related to removing a disconnected peer from
the server out of the peer handler since it is now protected by the
newly added peer state mutex.

This is a part of the overall effort to convert all of the code related
to updating and querying the server's peer state to synchronous code
that makes use of a separate mutex to protect it.