server: Make persistent peer removal synchronous.
This refactors the logic for manually removing persistent peers out of
the peer handler since it no longer needs to be plumbed through the
query channel.

This is a part of the overall effort to convert all of the code related
to updating and querying the server's peer state to synchronous code
that makes use of a separate mutex to protect it.