peer: Remove deprecated reject message support.
This removes code for handling deprecated reject messages which are no
longer supported by the network.

Since removing the handler definition itself would be a major module
version bump, it is not removed at this time and instead the deprecated
comment is updated to clearly communicate that the underlying message is
no longer valid and thus the handler will no longer ever be invoked.