peer: Require protocol v9 (removed reject msg).
This bumps the minimum required protocol version for peers to the
version that no longer supports the reject message now that a new
consensus change has taken place and all older nodes that might use it
are no longer able to follow the chain anyway.

It also adds a test to ensure peers with older protocol versions are
disconnected as intended.