wire: Remove deprecated reject message support.
This removes code for accepting reject messages which are no longer
supported by the network.

Since removing the type definition itself would be a major module
version bump, it is not removed at this time and instead the ability to
recognize the message is removed so that any attempts to read the
message will result in a protocol error.

The type and associated test code can be removed on a major module
version bump.