wire: Prepare v1.7.0.
This updates the wire module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
wire/v1.7.0.

The updated direct dependencies in this commit are as follows:

- lukechampine.com/blake3@v1.3.0

The full list of updated direct dependencies since the previous
wire/v1.6.0 release are as follows:

- lukechampine.com/blake3@v1.3.0

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.