chaincfg: Prepare v3.2.1.
This updates the chaincfg module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
chaincfg/v3.2.1.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/wire@v1.7.0

The updated indirect dependencies in this commit are as follows:

- lukechampine.com/blake3@v1.3.0

The full list of updated direct dependencies since the previous
chaincfg/v3.2.0 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.