addrmgr: Prepare v2.0.3.
This updates the addrmgr module dependencies and serves as a base for
addrmgr/v2.0.3.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/wire@v1.7.0

The updated indirect dependencies in this commit are as follows:

- lukechampine.com/blake3@v1.3.0

The full list of updated direct and indirect dependencies since the
previous addrmgr/v2.0.2 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.