connmgr: Prepare v3.1.2.
This updates the connmgr module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
connmgr/v3.1.2.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/wire@v1.7.0

The updated indirect dependencies in this commit are as follows:

- lukechampine.com/blake3@v1.3.0

The full list of updated dependencies since the previous connmgr/v3.1.1
release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.