txscript: Prepare v4.1.1.
This updates the txscript module dependencies and serves as a base for
txscript/v4.1.1.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg@3.2.1
- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.3.0
- github.com/decred/dcrd/wire@v1.7.0

The updated indirect dependencies in this commit are as follows:

- lukechampine.com/blake3@v1.3.0

The full list of updated dependencies since the previous txscript/v4.1.0
release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.