hdkeychain: Prepare v3.1.2.
This updates the hdkeychain module dependencies and serves as a base for
hdkeychain/v3.1.2.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/v3@v3.2.1
- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.3.0
- github.com/decred/dcrd/txscript/v4@v4.1.1

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/wire@v1.7.0
- lukechampine.com/blake3@v1.3.0

The full list of updated direct and indirect dependencies since the
previous hdkeychain/v3.1.1 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.