mixing: Add mixpool package.
The mixpool package implements a memory pool of recently observed mix
messages. Similar to the transaction mempool, the mixpool allows these
messages to be temporarily stored in memory to be relayed through the P2P
network. It handles message acceptance, expiry, UTXO ownership proof checks,
and that previously referenced messages have also been accepted to the
mixpool.

The mixpool is designed with both full-node and wallet usage in mind,
providing all of these same acceptance rules to mixing wallets with the
exception of UTXO proof checks. For wallets, it also implements query
functions for messages matching compatible pairings and messages belong to
ongoing sessions.