peer: Prepare v3.1.0.
This updates the peer module dependencies and serves as a base for
peer/v3.1.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/txscript/v4@v4.1.1
- github.com/decred/dcrd/wire@v1.7.0

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.3.0

The full list of updated direct dependencies since the previous
peer/v3.0.1 release are the same as above.

The full list of updated indirect dependencies since the previous
peer/v3.0.1 release are as follows:

- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.3.0
- lukechampine.com/blake3@v1.3.0

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.