database: Prepare v3.0.2.
This updates the database module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
database/v3.0.2.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg@v3.2.1
- github.com/decred/dcrd/dcrutil/v4@v4.0.2
- github.com/decred/dcrd/wire@v1.7.0

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.3.0
- github.com/decred/dcrd/txscript/v4@v4.1.1
- golang.org/x/sys@v0.20.0
- lukechampine.com/blake3@v1.3.0

The full list of updated direct dependencies since the previous
database/v3.0.1 release are as follows:

- github.com/decred/dcrd/chaincfg@v3.2.1
- github.com/decred/dcrd/dcrutil/v4@v4.0.2
- github.com/decred/dcrd/wire@v1.7.0
- github.com/jessevdk/go-flags@v1.5.0

The full list of updated indirect dependencies since the previous
database/v3.0.1 release are as follows:

- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.3.0
- github.com/decred/dcrd/txscript/v4@v4.1.1
- golang.org/x/sys@v0.20.0
- lukechampine.com/blake3@v1.3.0

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.