config: Show usage message on invalid cli flags.
This fixes a bug where the error handling code was not matching the
intended behaviour described in its comment, and as a result the usage
message was not being displayed when an invalid config flag was passed.
dcrd was also exiting with code 0 instead of 1.

After fixing this bug the usage message was shown as expected and the
exit code was correct, but the config error was being printed twice -
once by dcrd itself, and once by the go-flags lib. This was
fixed by changing the options passed into go-flags.