gcs: Prepare v4.1.0.
This updates the gcs module dependencies, the copyright year in the
files modified since the previous release,  and serves as a base for
gcs/v4.1.0.

The full list of updated direct dependencies since the previous
gcs/v4.0.0 release are as follows:

- github.com/decred/dcrd/blockchain/stake/v5@v5.0.1
- github.com/decred/dcrd/chaincfg/v3@v3.2.1
- github.com/decred/dcrd/txscript/v4@v4.1.1
- github.com/decred/dcrd/wire@v1.7.0

The full list of updated indirect dependencies since the previous
gcs/v4.0.0 release are as follows:

- github.com/decred/dcrd/database/v3@v3.0.2
- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.3.0
- github.com/decred/dcrd/dcrutil/v4@v4.0.2
- lukechampine.com/blake3@v1.3.0

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.