rpcclient: Prepare v8.0.1.
This updates the rpcclient module dependencies and serves as a base for
rpcclient/v8.0.1.

The full list of updated direct dependencies since the previous
rpcclient/v8.0.0 release are as follows:

- github.com/decred/dcrd/dcrutil/v4@v4.0.2
- github.com/decred/dcrd/gcs/v4@v4.1.0
- github.com/decred/dcrd/rpc/jsonrpc/types/v4@v4.2.0
- github.com/decred/dcrd/txscript/v4@v4.1.1
- github.com/decred/dcrd/wire@v1.7.0
- github.com/gorilla/websocket@v1.5.1

The full list of updated indirect dependencies since the previous
rpcclient/v8.0.0 release are as follows:

- github.com/decred/dcrd/blockchain/stake/v5@v5.0.1
- github.com/decred/dcrd/chaincfg/v3@v3.2.1
- github.com/decred/dcrd/database/v3@v3.0.2
- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.3.0
- golang.org/x/net@v0.17.0
- lukechampine.com/blake3@v1.3.0

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a
result.