[release-v2.0] mixpool: Wrap non-bannable errors with RuleError
To reduce spurious error logs when invalid messages are received off the
network, the RuleError type is expanded to cover submission errors that should
not necessarily result in an instant permanent ban.

A new function IsBannable is added to check whether the error condition is
still such that it should be treated as bannable.