netsync: Request new headers on rejected tip.
This modifies the sync manager to immediately request new headers from
all peers in the event the block associated with the best target header
is rejected.  This allows alternative blocks to be downloaded more
quickly without first having to wait for new blocks.

While here, it also allows a couple of blocks of leeway when returning
the manager to an unsynced state.