[release-v2.0] dcrjson,rpcserver,types: Add getmixmessage
This adds a new RPC named getmixmessage to the JSON-RPC server which queries
the current mixpool for a mixing message by its hash.

This RPC will be used in later wallet releases to fetch unknown pair request
messages when a run-0 key exchange is rejected from the wallet's mixpool for
referencing an unknown PR by the same identity.  This is also how initial PR
sync occurs during the warmup period, replacing the earlier use of
getmixpairrequests.