mixclient: Do not remove PRs from failed runs
Pair request messages were being unintentionally discarded from the mixpool
even on sessions that did not run to completion.  This was occurring under two
separate situations:

The first situation occured if the context passed to the mixpool's Receive
method while receiving KEs was cancelled.  A last-minute change in semantics
of the error return result of the Receive method during review of the mixpool
package resulted in this method no longer returning the non-nil context error.
Most of the client was updated for this change in semantics, but the KE
receiving is special and was not handling the change properly.  This resulted
in the run method returning with a nil error, and the PRs from the incompleted
session were removed like it was actually completed.

The second situation is a longstanding bug unrelated to the recent mixpool
review.  If a run is aborted due to the next epoch about to start, this was
also causing the run method to return with a nil error, and the PRs were
likewise incorrectly removed.

In both of these situations, peers were still trying to mix using these PRs,
but other peers on the network would not accept their later KEs after the PRs
had been removed.

Debug logging has been improved substantially as part of debugging these
issues, and in particular, the session ID and the coinjoin transaction hash
are now both logged after a mix is successful and the PRs are finally (and
correctly) removed from the mixpool.