server: Add dynamic profile server infrastructure.
This adds infrastructure to support dynamically starting and stopping
the profile server and updates the code that allows the profile to be
started via the --profile command line / "profile=" config file option
to use the new infrastructure such that it retains the existing
behavior.

As a part of the changes, it expands the capabilities of the
aforementioned profile options to allow specifying an interface name so
it is more consistent with the other options that deal with listeners.

An overview of the changes are:

- Refactors profile listening port normalization and validation
  - Helps ensure consistent behavior regardless of how the profile
    server listen address is obtained
- Introduces a profile server type with methods to dynamically start,
  stop, and query the profile server
  - Includes a new flag to control whether or not non loopback addresses
    are allowed for future dynamic control
- Adds support for resolving interface names for the profile server
  listen address to make the behavior more consistent