netsync: Split header sync from block sync.
This refactors the code that deals with starting synchronization with
the sync peer to split out the pieces specific to the initial header
sync from the chain data sync.  Along the way it removes some duplicate
checks that are no longer necessary.

The primary motivation is to help pave the way for eventually performing
parallel block downloads, but it also has the added benefit of making
the logic a bit easier to follow, so it is useful on its own.