[release-v2.0] mixclient: Unexport ErrExpired
In future releases, expired PRs will be handled by creating a new PR for the
existing client call, so there will be no need to detect or expose this error.

While here, add additional context describing the PR which expired.  This info
is not terribly useful to the caller, but it is only temporary until expired
PRs are eventually recreated with a higher expiry.