mixpool: Remove unused exported methods
The MixPRs method has been modified to remove the query parameter, as it was
only ever called with nil.

The RPC server no longer needs access to directly call RemoveConfirmedMixes.
This same operation is already (and incorrectly) done by MixPRs, and it will
need the eventual fix.

One exception: we are keeping RemoveConfirmedMixes which is still not used
anywhere.  This method behaves similarly to RemoveSpentPRs but removes mixes
that completed with a confirmed coinjoin transaction matching any of the
provided hashes.  This may be used later by wallet to remove mixes that
include transaction hashes in the merkle tree, while dcrd can just provide the
mined transactions themselves and remove anything that double spends a PR
UTXO.