uniform: Add internal uniform random package
Code adapted from dcrwallet's internal/uniformprng package, but modified to
take any reader as a randomness source.  This can be a CSPRNG, or
crypto/rand.Reader.

The random reader is required to never panic.  This will not be a concern when
using crypto/rand on any modern operating system which provides non-blocking
kernel entropy through a system call.