peer: Use container/lru module.
This updates the peer known inventory and sent nonces LRU caches to use
the new container/lru module.

It also uses the expiration functionality of the new module to impose an
expiration time of 15 minutes on the known inventory entries.  This will
allow the possibility of the cache shrinking over time in periods of low
activity versus the current behavior where it will eventually reach the
limit and stay there indefinitely.